Moved the text pivot to the left again to circumvent rendering bugs in both IE 11 and Edge.
It seems like both of Microsoft's browsers are incapable of correctly
calculating the center of text (which may be related to our text
bounds) which causes the text to shift ever so slightly to the right
within some badges. Specifying the position normally circumvents this
issues (but introduces some slight shifting in a limited set of badges
due to the inaccuracy of the server side positioning).